Move evaluation state logic to EvaluationStateHelper
